\name{CheckParamRange}
\alias{CheckParamRange}

\title{Check of the Parameters' Ranges}
\description{
  Subroutine called by the fitting procedures.
  The procedure controls the range of the correlation's parameters.
}
\usage{
CheckParamRange(param)
} 
\arguments{
  \item{param}{Numeric; a vector of correlation's parameters.}
}

\seealso{\code{\link{FitComposite}}}

\author{Simone Padoan, \email{simone.padoan@epfl.ch},
  \url{http://eflum.epfl.ch/people/padoan.en.php}.}

\keyword{Composite}
