\name{NuisanceParam}
\alias{NuisanceParam}

\title{Lists the Nuisance Parameters of a Random Field}
\description{
  Subroutine called by InitParam and other procedures.
  The procedure returns a list with the nuisance parameters of a given
  random field model.
}
\usage{
NuisanceParam(model)
}
\arguments{
  \item{model}{String; the name of a random field.}
}

\seealso{\code{\link{FitComposite}}}

\author{Simone Padoan, \email{simone.padoan@stat.unipd.it},
  \url{http://homes.stat.unipd.it/padoan}.}


\keyword{Composite}
