\name{usrain}
\alias{usrain}
\title{Annual Maximum Rainfalls in U.S.}
\usage{data(usrain)}
\description{
  A (\eqn{91 \times 46}{91 x 46})-matrix containing annual maximum of daily
  rainfalls, in millimeters, for 91 years at 46 gauging stations in U.S.
}
\format{A numerical matrix containing 4186 observations.}
\source{
  Padoan, S. A. Ribatet, M. and Sisson, S. A. (2010)
  Likelihood-Based Inference for Max-Stable Processes.
  \emph{Journal of the American Statistical Association, Theory & Methods}, \bold{105}, 263--277.
}
\keyword{datasets}
