\name{us.coords}
\alias{us.coords}
\title{Gauging Stations of the Annual Maximum Rainfalls in U.S.}
\usage{data(usrain)}
\description{
  A (\eqn{46 \times 3}{46 x 3})-matrix containing the coordinates
  of the 46 gauging stations in U.S. where the daily precipitation are
  recorded.
}
\format{A numerical matrix containing the longitude, the latitude and
  the
  altitude of the gauging stations.}
\source{
  Padoan, S. A. Ribatet, M. and Sisson, S. A. (2010)
  Likelihood-Based Inference for Max-Stable Processes.
  \emph{Journal of the American Statistical Association, Theory & Methods}, \bold{105}, 263--277.
}
\keyword{datasets}
