\name{CheckCorrModel}
\alias{CheckCorrModel}

\title{Checking Correlation Model}
\description{
  Subroutine called by InitParam.
  The procedure controls if the correlation model
  inserted is correct.
}
\usage{
CheckCorrModel(corrmodel)
}
\arguments{
  \item{corrmodel}{String; the name of a correlation model, for the
    description see \code{\link{Covmatrix}}.}
}

\seealso{\code{\link{FitComposite}}}

\author{Simone Padoan, \email{simone.padoan@unibocconi.it},
  \url{http://faculty.unibocconi.it/simonepadoan};
  Moreno Bevilacqua, \email{moreno.bevilacqua@uv.cl},
   \url{http://staff.deuv.cl/bevilacqua/}.
  }

\keyword{Composite}
