\name{CorrelationParam}
\alias{CorrelationParam}

\title{Lists the Parameters of the Correlation function}
\description{
  Subroutine called by DetectParam and InitParam.
  The procedure returns the list of the parameter for a given
  correlation model.
}
\usage{
CorrelationParam(corrmodel)
} 
\arguments{
  \item{corrmodel}{String; the name of a correlation model.}
}

\seealso{\code{\link{FitComposite}}}

\author{Simone Padoan, \email{simone.padoan@epfl.ch},
  \url{http://eflum.epfl.ch/people/padoan.en.php}.}

\keyword{Composite}
